requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
Django==3.2.4
django-cors-headers==3.11.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.2.1
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.0.1
pymongo==3.12.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scipy==1.5.4
six==1.15.0
sqlparse==0.2.4
urllib3==1.26.6
